requirements.txt
python-dotenv
openai>=1.12.0
openai-agents>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
requests>=2.31.0
beautifulsoup4>=4.12.2
aiohttp>=3.8.6
websockets>=12.0
python-multipart>=0.0.6